Refactor metagraph into it's own package
