+++ src/main/java/org/lab41/dendrite/web/controller/JobController.java
@@ -1,10 +1,10 @@
 package org.lab41.dendrite.web.controller;
 
-import org.lab41.dendrite.models.GraphMetadata;
-import org.lab41.dendrite.models.JobMetadata;
-import org.lab41.dendrite.models.ProjectMetadata;
+import org.lab41.dendrite.metagraph.MetaGraphTx;
+import org.lab41.dendrite.metagraph.models.GraphMetadata;
+import org.lab41.dendrite.metagraph.models.JobMetadata;
+import org.lab41.dendrite.metagraph.models.ProjectMetadata;
 import org.lab41.dendrite.services.MetadataService;
-import org.lab41.dendrite.services.MetadataTx;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,8 +12,6 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.util.UriComponents;
-import org.springframework.web.util.UriComponentsBuilder;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -30,7 +28,7 @@ public class JobController {
     @RequestMapping(value = "/jobs", method = RequestMethod.GET)
     public ResponseEntity<Map<String, Object>> getJobs() {
 
-        MetadataTx tx = metadataService.newTransaction();
+        MetaGraphTx tx = metadataService.newTransaction();
 
         List<Map<String, Object>> jobs = new ArrayList<>();
         for (JobMetadata jobMetadata: tx.getJobs()) {
@@ -50,7 +48,7 @@ public class JobController {
     public ResponseEntity<Map<String, Object>> getJob(@PathVariable String jobId) {
 
         Map<String, Object> response = new HashMap<>();
-        MetadataTx tx = metadataService.newTransaction();
+        MetaGraphTx tx = metadataService.newTransaction();
         JobMetadata jobMetadata = tx.getJob(jobId);
 
         if (jobMetadata == null) {
@@ -72,7 +70,7 @@ public class JobController {
     public ResponseEntity<Map<String, Object>> getJobs(@PathVariable String projectId) {
 
         Map<String, Object> response = new HashMap<>();
-        MetadataTx tx = metadataService.newTransaction();
+        MetaGraphTx tx = metadataService.newTransaction();
         ProjectMetadata project = tx.getProject(projectId);
 
         if (project == null) {
@@ -99,7 +97,7 @@ public class JobController {
     public ResponseEntity<Map<String, Object>> deleteJob(@PathVariable String jobId) {
 
         Map<String, Object> response = new HashMap<>();
-        MetadataTx tx = metadataService.newTransaction();
+        MetaGraphTx tx = metadataService.newTransaction();
         JobMetadata jobMetadata = tx.getJob(jobId);
 
         if (jobMetadata == null) {
